[Edit] mood to int and refactor

+++ app/src/main/java/com/kstudio/diarymylife/widgets/CustomCard.kt
@@ -22,7 +22,7 @@ class CustomCard @JvmOverloads constructor(
         cardIcon?.let { icon.setImageResource(it) }
     }
 
-    fun setValue(cardValue: Int) = with(binding){
+    fun setValue(cardValue: Long) = with(binding){
         value.text = cardValue.toString()
     }
 }
